refactor 기업 템플릿 Command 패턴 적용

+++ backend/src/main/java/com/beyond/specguard/companytemplate/model/service/CompanyTemplateServiceImpl.java
@@ -3,7 +3,10 @@ package com.beyond.specguard.companytemplate.model.service;
 import com.beyond.specguard.common.exception.CustomException;
 import com.beyond.specguard.companytemplate.exception.ErrorCode.CompanyTemplateErrorCode;
 import com.beyond.specguard.companytemplate.model.dto.command.CreateCompanyTemplateFieldCommand;
+import com.beyond.specguard.companytemplate.model.dto.command.UpdateTemplateBasicCommand;
 import com.beyond.specguard.companytemplate.model.dto.command.UpdateTemplateDetailCommand;
+import com.beyond.specguard.companytemplate.model.dto.request.CompanyTemplateBasicRequestDto;
+import com.beyond.specguard.companytemplate.model.dto.request.CompanyTemplateDetailRequestDto;
 import com.beyond.specguard.companytemplate.model.dto.request.TemplateFieldRequestDto;
 import com.beyond.specguard.companytemplate.model.entity.CompanyTemplate;
 import com.beyond.specguard.companytemplate.model.entity.CompanyTemplateField;
@@ -11,6 +14,7 @@ import com.beyond.specguard.companytemplate.model.repository.CompanyTemplateRepo
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -26,6 +30,8 @@ public class CompanyTemplateServiceImpl implements CompanyTemplateService {
     private final CompanyTemplateFieldService companyTemplateFieldService;
 
     @Override
+    // 부모 Transaction에 참여하지 않음
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public CompanyTemplate getCompanyTemplate(UUID templateId) {
         return companyTemplateRepository.findById(templateId)
                 .orElseThrow(() -> new CustomException(CompanyTemplateErrorCode.TEMPLATE_NOT_FOUND));
@@ -34,12 +40,25 @@ public class CompanyTemplateServiceImpl implements CompanyTemplateService {
     @Override
     @Transactional
     public void deleteTemplate(UUID templateId) {
+        // id가 존재하지 않으면 예외 던지기
+        if (!companyTemplateRepository.existsById(templateId)) {
+            throw new CustomException(CompanyTemplateErrorCode.TEMPLATE_NOT_FOUND);
+        }
+
+        // template 삭제
         companyTemplateRepository.deleteById(templateId);
+
+        // fk 제약이 없으므로  templateField 수동 삭제
+        companyTemplateFieldService.deleteFields(templateId);
     }
 
     @Override
     @Transactional
-    public CompanyTemplate updateTemplate(CompanyTemplate companyTemplate) {
+    public CompanyTemplate updateBasic(UpdateTemplateBasicCommand command) {
+        CompanyTemplate companyTemplate = getCompanyTemplate(command.templateId());
+
+        companyTemplate.update(command.requestDto());
+
         return companyTemplateRepository.save(companyTemplate);
     }
 
@@ -58,7 +77,7 @@ public class CompanyTemplateServiceImpl implements CompanyTemplateService {
         List<CompanyTemplateField> existingFields = companyTemplateFieldService.getFields(templateId);
 
         // 2. 요청 필드 id → dto 매핑
-        Map<UUID, TemplateFieldRequestDto> dtoMap = command.requestDto().getFields().stream()
+        Map<UUID, TemplateFieldRequestDto> dtoMap = command.requestDto().fields().stream()
                 .filter(f -> f.getId() != null)
                 .collect(Collectors.toMap(TemplateFieldRequestDto::getId, f -> f));
 
@@ -74,7 +93,7 @@ public class CompanyTemplateServiceImpl implements CompanyTemplateService {
         }
 
         // 4. 신규 생성
-        command.requestDto().getFields().stream()
+        command.requestDto().fields().stream()
                 .filter(f -> f.getId() == null)
                 .forEach(newDto -> {
                     CreateCompanyTemplateFieldCommand companyTemplateFieldCommand = new CreateCompanyTemplateFieldCommand(updatedTemplate, newDto);
@@ -86,22 +105,36 @@ public class CompanyTemplateServiceImpl implements CompanyTemplateService {
 
     @Override
     @Transactional
-    public CompanyTemplate createTemplate(CompanyTemplate template) {
-        CompanyTemplate companyTemplate = template;
+    public CompanyTemplate createDetailTemplate(CompanyTemplateDetailRequestDto requestDto) {
+        UUID templateId = requestDto.detailDto().getTemplateId();
 
-        if (template.getId() == null) {
-            // 신규 생성 케이스: ID가 없으므로 save() → INSERT 발생
-            companyTemplate = companyTemplateRepository.save(template);
-        } else {
-            // 업데이트 케이스: ID가 존재하는 경우
-            if (companyTemplateRepository.existsById(template.getId())) {
-                template.setActive(true);
-                companyTemplate = companyTemplateRepository.save(template);
-            } else {
-                // ID는 있는데 DB에 존재하지 않는 경우 → 예외 처리
-                throw new CustomException(CompanyTemplateErrorCode.TEMPLATE_NOT_FOUND);
-            }
-        }
-        return companyTemplate;
+        // 1. CompanyTemplate Detail 부분 반영해서 업데이트
+        CompanyTemplate companyTemplate = getCompanyTemplate(templateId);
+
+        // 1.1 사용 완료 상태 True로 하기
+        companyTemplate.setStatusActive();
+
+        CompanyTemplate createdTemplate = companyTemplateRepository.save(companyTemplate);
+
+        // 2. Field 저장
+        // 2.1 CreateCompanyTemplateFieldCommand로 변경
+        List<CreateCompanyTemplateFieldCommand> commands = requestDto.fields()
+                .stream()
+                .map(dto -> new CreateCompanyTemplateFieldCommand(createdTemplate, dto))
+                .toList();
+
+        // 2.2 Field 저장
+        companyTemplateFieldService.createFields(commands);
+
+        return createdTemplate;
+    }
+
+    @Override
+    public CompanyTemplate createBasicTemplate(CompanyTemplateBasicRequestDto basicRequestDto) {
+        // BasicRequestDto 에서 받은 정보를 토대로 entity 생성
+        // 그런데, not null 제약조건 있는 것은 기본 값 삽입하기.
+        CompanyTemplate companyTemplate = basicRequestDto.toEntity();
+
+        return companyTemplateRepository.save(companyTemplate);
     }
 }
